+++ nummus/controllers/__init__.py
@@ -12,7 +12,6 @@ from nummus.controllers import (
     auth,
     budgeting,
     common,
-    dashboard,
     emergency_fund,
     health,
     import_file,
@@ -43,7 +42,6 @@ def add_routes(app: flask.Flask, *, debug: bool) -> None:
         auth,
         budgeting,
         common,
-        dashboard,
         emergency_fund,
         health,
         import_file,
