+++ pyproject.toml
@@ -98,10 +98,9 @@ command_line = "-m pytest"
 branch = true
 source = ["nummus"]
 omit = [
-  "nummus/__main__.py",         # Not capture by coverage
-  "nummus/version.py",          # Auto generated by setuptools-scm
-  "nummus/controllers/base.py", # Currently just a type definition
-  "nummus/commands/web.py",     # Just a wrapper on web.Server.run() and difficult to mock
+  "nummus/__main__.py",     # Not capture by coverage
+  "nummus/version.py",      # Auto generated by setuptools-scm
+  "nummus/commands/web.py", # Just a wrapper on web.Server.run() and difficult to mock
   # TODO (WattsUp): Migrate tests
   "nummus/controllers/*",
   "nummus/web/*",
